Moving logic to create different types of Analytics events into the Analytics module
